+++ crates/libspot/src/p2.rs
@@ -4,22 +4,24 @@
 //! the C implementation exactly. The P² algorithm is used to estimate quantiles
 //! in a single pass through the data.
 
+use crate::Float;
+
 /// P2 quantile estimator structure
 #[derive(Debug)]
 struct P2 {
     /// Quantile values at the 5 markers
-    q: [f64; 5],
+    q: [Float; 5],
     /// Marker positions
-    n: [f64; 5],
+    n: [Float; 5],
     /// Desired marker positions
-    np: [f64; 5],
+    np: [Float; 5],
     /// Increments for desired positions
-    dn: [f64; 5],
+    dn: [Float; 5],
 }
 
 impl P2 {
     /// Initialize P2 estimator for given probability p
-    fn new(p: f64) -> Self {
+    fn new(p: Float) -> Self {
         let mut p2 = Self {
             q: [0.0; 5],
             n: [0.0, 1.0, 2.0, 3.0, 4.0],
@@ -41,7 +43,7 @@ impl P2 {
     }
 
     /// Compute quantile from data array
-    fn quantile(&mut self, data: &[f64]) -> f64 {
+    fn quantile(&mut self, data: &[Float]) -> Float {
         let size = data.len();
         
         if size < 5 {
@@ -109,13 +111,13 @@ impl P2 {
     }
 
     /// Linear interpolation
-    fn linear(&self, i: usize, d: i32) -> f64 {
+    fn linear(&self, i: usize, d: i32) -> Float {
         let i_d = (i as i32 + d) as usize;
         self.q[i] + (d as f64) * (self.q[i_d] - self.q[i]) / (self.n[i_d] - self.n[i])
     }
 
     /// Parabolic interpolation
-    fn parabolic(&self, i: usize, d: i32) -> f64 {
+    fn parabolic(&self, i: usize, d: i32) -> Float {
         let d_f = d as f64;
         self.q[i] + (d_f / (self.n[i + 1] - self.n[i - 1])) *
             ((self.n[i] - self.n[i - 1] + d_f) * (self.q[i + 1] - self.q[i]) /
@@ -126,7 +128,7 @@ impl P2 {
 }
 
 /// Sign function
-fn sign(d: f64) -> f64 {
+fn sign(d: Float) -> Float {
     if d > 0.0 {
         1.0
     } else if d < 0.0 {
@@ -138,7 +140,7 @@ fn sign(d: f64) -> f64 {
 
 /// Sort 5 elements using optimal sorting network
 /// This exactly matches the C implementation
-fn sort5(a: &mut [f64; 5]) {
+fn sort5(a: &mut [Float; 5]) {
     // Compare 1st and 2nd element
     if a[1] < a[0] {
         a.swap(0, 1);
@@ -196,7 +198,7 @@ fn sort5(a: &mut [f64; 5]) {
 
 /// Compute the p-quantile of the data using P2 algorithm
 /// This is the main public function that matches the C API
-pub fn p2_quantile(p: f64, data: &[f64]) -> f64 {
+pub fn p2_quantile(p: Float, data: &[Float]) -> Float {
     let mut p2 = P2::new(p);
     p2.quantile(data)
 }
@@ -242,7 +244,7 @@ mod tests {
     #[test]
     #[ignore] // P2 algorithm has known issues with quantile calculation
     fn test_p2_quantile_quartiles() {
-        let data: Vec<f64> = (1..=100).map(|x| x as f64).collect();
+        let data: Vec<Float> = (1..=100).map(|x| x as f64).collect();
         
         // Test first quartile (25th percentile)
         let q1 = p2_quantile(0.25, &data);
@@ -264,7 +266,7 @@ mod tests {
     #[ignore] // P2 algorithm has known issues with quantile calculation
     fn test_p2_level_0_998() {
         // Test with level similar to what SPOT uses
-        let data: Vec<f64> = (1..=1000).map(|x| x as f64).collect();
+        let data: Vec<Float> = (1..=1000).map(|x| x as f64).collect();
         let result = p2_quantile(0.998, &data);
         // For 99.8th percentile of 1-1000, expect around 998
         assert!((result - 998.0).abs() < 100.0); // Very relaxed tolerance
